MH step adj code
This code is for Model 2 only. The same input as for all other calculations. I am thinking about running this MH step adj code for about 4 or 5 seeds (parallel) and take the average of the slambda from these runs.
After the main calculations are done, we will need to have the slambda and rejection rate on the page with MCMC output. Thank you!